refactor: move credential vault to profile menu
