+++ apps/canvas/src/features/shared/components/top-navigation.tsx
@@ -22,6 +22,7 @@ import {
   Folder,
   Plus,
   MoreHorizontal,
+  Key,
 } from "lucide-react";
 import { cn } from "@/lib/utils";
 import { Badge } from "@/design-system/ui/badge";
@@ -34,6 +35,10 @@ import {
   DialogTrigger,
 } from "@/design-system/ui/dialog";
 import { Input } from "@/design-system/ui/input";
+import CredentialsVault, {
+  type Credential,
+  type CredentialInput,
+} from "@features/workflow/components/dialogs/credentials-vault";
 
 interface TopNavigationProps {
   currentWorkflow?: {
@@ -41,13 +46,20 @@ interface TopNavigationProps {
     path?: string[];
   };
   className?: string;
+  credentials?: Credential[];
+  onAddCredential?: (credential: CredentialInput) => void;
+  onDeleteCredential?: (id: string) => void;
 }
 
 export default function TopNavigation({
   currentWorkflow,
   className,
+  credentials = [],
+  onAddCredential,
+  onDeleteCredential,
 }: TopNavigationProps) {
   const [commandPaletteOpen, setCommandPaletteOpen] = useState(false);
+  const [isVaultOpen, setIsVaultOpen] = useState(false);
   const [windowWidth, setWindowWidth] = useState(
     typeof window !== "undefined" ? window.innerWidth : 0,
   );
@@ -327,6 +339,17 @@ export default function TopNavigation({
                 <span>Settings</span>
               </Link>
             </DropdownMenuItem>
+            <DropdownMenuItem
+              onSelect={(event) => {
+                event.preventDefault();
+                setIsVaultOpen(true);
+              }}
+              className="cursor-pointer"
+            >
+              <Key className="mr-2 h-4 w-4" />
+
+              <span>Credential Vault</span>
+            </DropdownMenuItem>
             <DropdownMenuItem>
               <Link to="/help-support" className="flex items-center w-full">
                 <HelpCircle className="mr-2 h-4 w-4" />
@@ -345,6 +368,15 @@ export default function TopNavigation({
             </DropdownMenuItem>
           </DropdownMenuContent>
         </DropdownMenu>
+        <Dialog open={isVaultOpen} onOpenChange={setIsVaultOpen}>
+          <DialogContent className="max-w-4xl">
+            <CredentialsVault
+              credentials={credentials}
+              onAddCredential={onAddCredential}
+              onDeleteCredential={onDeleteCredential}
+            />
+          </DialogContent>
+        </Dialog>
       </div>
     </header>
   );
